c2-samples/src/main/java/aut/utcluj/isp/ex1/Equipment.java
package aut.utcluj.isp.ex1;

/**
 * @author stefan
 */
public class Equipment {
    private String name;
    private String serialNumber;

    public Equipment(String serialNumber) {
        this.name = "NONE";
        throw new UnsupportedOperationException("Not supported yet.");
    }

    public Equipment(String name, String serialNumber) {
        throw new UnsupportedOperationException("Not supported yet.");
    }

    public String getName() {
        return name;
    }

    public String getSerialNumber() {
        return serialNumber;
    }
}
